+++ lib/src/web/datadog_web_rum.dart
@@ -43,11 +43,12 @@ class DatadogWebRum {
       case 'rumStopView':
         return false;
       case 'setUserInfo':
-        dd_rum.setUser(dd_rum.UserOptions(
-          email: call.arguments['email'],
-          id: call.arguments['id'],
-          name: call.arguments['name'],
-        ));
+        dd_rum.setUser(jsify({
+          'email': call.arguments['email'],
+          'id': call.arguments['id'],
+          'name': call.arguments['name'],
+          ...?call.arguments['extraInfo']
+        }));
         return true;
       default:
         return null;