+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Remove `Platform` bug for web
+* Acknowledge `extraInfo` arguments for Flutter Web users (#76)
 
 ## 1.5.2